Prevents deleting the admin user

+++ include/php/pages/admin/deletedomain.php
@@ -2,21 +2,25 @@
 
 $id = $db->escape_string($_GET['id']);
 
+//Load user data from DB
+$sql = "SELECT `".DBC_DOMAINS_DOMAIN."` FROM `".DBT_DOMAINS."` WHERE `".DBC_DOMAINS_ID."` = '$id' LIMIT 1;";
+
+if(!$result = $db->query($sql)){
+	die('There was an error running the query [' . $db->error . ']');
+}
+
+while($row = $result->fetch_assoc()){
+	$domain = $row[DBC_DOMAINS_DOMAIN];
+}
+
+// Delete domain
 if(isset($_POST['confirm'])){
 	$confirm = $_POST['confirm'];
 	
 	if($confirm === "yes"){
-		$sql = "SELECT `".DBC_DOMAINS_DOMAIN."` FROM `".DBT_DOMAINS."` WHERE `".DBC_DOMAINS_ID."` = '$id' LIMIT 1;";
-			
-		if(!$result = $db->query($sql)){
-			die('There was an error running the query [' . $db->error . ']');
-		}
-		
-		else{	
-			while($row = $result->fetch_assoc()){
-				$domain = $row[DBC_DOMAINS_DOMAIN];
-			}
-			
+		// Check if admin domain is affected
+		$admin_domain = explode("@", ADMIN_EMAIL)[1];
+		if($admin_domain !== $domain){				
 			$sql = "DELETE FROM `".DBT_DOMAINS."` WHERE `".DBC_DOMAINS_ID."` = '$id'";
 				
 			if(!$result = $db->query($sql)){
@@ -34,23 +38,13 @@ if(isset($_POST['confirm'])){
 				}
 			}
 		}
+		else{
+			header("Location: ".FRONTEND_BASE_PATH."admin/listdomains/?adm_del=1");
+		}
 	}
 	
 	else{
 		header("Location: ".FRONTEND_BASE_PATH."admin/listdomains/");
-	}
-}
-
-else{
-	//Load user data from DB
-	$sql = "SELECT `".DBC_DOMAINS_DOMAIN."` FROM `".DBT_DOMAINS."` WHERE `".DBC_DOMAINS_ID."` = '$id' LIMIT 1;";
-	
-	if(!$result = $db->query($sql)){
-		die('There was an error running the query [' . $db->error . ']');
-	}
-	
-	while($row = $result->fetch_assoc()){
-		$domain = $row[DBC_DOMAINS_DOMAIN];
 	}
 }
 ?>
