+++ include/php/pages/admin/deleteuser.php
@@ -2,17 +2,39 @@
 
 $id = $db->escape_string($_GET['id']);
 
+//Load user data from DB
+$sql = "SELECT `".DBC_USERS_USERNAME."`, `".DBC_USERS_DOMAIN."` FROM `".DBT_USERS."` WHERE `".DBC_USERS_ID."` = '$id' LIMIT 1;";
+
+if(!$result = $db->query($sql)){
+	die('There was an error running the query [' . $db->error . ']');
+}
+
+while($row = $result->fetch_assoc()){
+	$username = $row[DBC_USERS_USERNAME];
+	$domain = $row[DBC_USERS_DOMAIN];
+}
+
+$mailaddress = $username."@".$domain;
+
+
+// Delete user
 if(isset($_POST['confirm'])){
 	$confirm = $_POST['confirm'];
 	
 	if($confirm === "yes"){
-		$sql = "DELETE FROM `".DBT_USERS."` WHERE `".DBC_USERS_ID."` = '$id'";
-			
-		if(!$result = $db->query($sql)){
-			die('There was an error running the query [' . $db->error . ']');
+		if($mailaddress !== ADMIN_EMAIL){
+			$sql = "DELETE FROM `".DBT_USERS."` WHERE `".DBC_USERS_ID."` = '$id'";
+				
+			if(!$result = $db->query($sql)){
+				die('There was an error running the query [' . $db->error . ']');
+			}
+			else{
+				header("Location: ".FRONTEND_BASE_PATH."admin/listusers/?deleted=1");
+			}
 		}
 		else{
-			header("Location: ".FRONTEND_BASE_PATH."admin/listusers/?deleted=1");
+			// Admin tries to delete himself. WTH.
+			header("Location: ".FRONTEND_BASE_PATH."admin/listusers/?adm_del=1");
 		}
 	}
 	
@@ -21,21 +43,6 @@ if(isset($_POST['confirm'])){
 	}
 }
 
-else{
-	//Load user data from DB
-	$sql = "SELECT `".DBC_USERS_USERNAME."`, `".DBC_USERS_DOMAIN."` FROM `".DBT_USERS."` WHERE `".DBC_USERS_ID."` = '$id' LIMIT 1;";
-	
-	if(!$result = $db->query($sql)){
-		die('There was an error running the query [' . $db->error . ']');
-	}
-	
-	while($row = $result->fetch_assoc()){
-		$username = $row[DBC_USERS_USERNAME];
-		$domain = $row[DBC_USERS_DOMAIN];
-	}
-	
-	$mailaddress = $username."@".$domain;
-}
 ?>
 
 <h1>Delete user "<?php echo $mailaddress ?>"?</h1>
